--- src/test/java/com/scc/health/HealthCheckIT.java
@@ -1,7 +1,0 @@
-package com.scc.health;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class HealthCheckIT extends HealthCheckTest {
-}
