+++ src/test/java/com/scc/keycloak/NativePolicyEnforcerIT.java
@@ -1,0 +1,10 @@
+package com.scc.keycloak;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+/**
+ * @author <a href="mailto:psilva@redhat.com">Pedro Igor</a>
+ */
+@NativeImageTest
+public class NativePolicyEnforcerIT extends PolicyEnforcerTest {
+}
