+++ src/test/java/com/scc/keycloak/PolicyEnforcerTest.java
@@ -1,5 +1,10 @@
 package com.scc.keycloak;
 
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.CoreMatchers.anyOf;
+import static org.hamcrest.CoreMatchers.everyItem;
+import static org.hamcrest.CoreMatchers.is;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.keycloak.representations.AccessTokenResponse;
@@ -14,18 +19,28 @@ public class PolicyEnforcerTest {
     private static final String KEYCLOAK_SERVER_URL = System.getProperty("keycloak.url", "http://localhost:8180/auth");
     private static final String KEYCLOAK_REALM = "quarkus";
 
+    /*
+    @Test
+    public void testHealthCheck() {
+        given()
+                .when()
+                .get("/public/health/live")
+                .then()
+                .statusCode(200)
+                .body("checks.name", everyItem(anyOf(
+                        is("WS is alive !"),
+                        is("Health check with data"))))
+       ;
+    }
+    */
     @Test
     public void testAccessUserResource() {
         RestAssured.given().auth().oauth2(getAccessToken("alice"))
-                .when().get("/api/users/me")
-                .then()
-                .statusCode(200);
-        RestAssured.given().auth().oauth2(getAccessToken("jdoe"))
-                .when().get("/api/users/me")
+                .when().get("/api/pedigrees/token/JVP685")
                 .then()
                 .statusCode(200);
     }
-
+    /*
     @Test
     public void testAccessAdminResource() {
         RestAssured.given().auth().oauth2(getAccessToken("alice"))
@@ -41,15 +56,16 @@ public class PolicyEnforcerTest {
                 .then()
                 .statusCode(200);
     }
-
+    */
+    /*
     @Test
     public void testPublicResource() {
         RestAssured.given()
-                .when().get("/api/public")
+                .when().get("/public")
                 .then()
-                .statusCode(204);
+                .statusCode(200);
     }
-
+    */
     private String getAccessToken(String userName) {
         return RestAssured
                 .given()
