+++ src/test/resources/application.properties
@@ -1,9 +1,31 @@
+# Configuration file
+quarkus.oidc.auth-server-url=http://localhost:8180/auth/realms/quarkus
+quarkus.oidc.client-id=backend-service
+quarkus.oidc.credentials.secret=secret
+quarkus.http.cors=true
+
+# Enable Policy Enforcement
+quarkus.keycloak.policy-enforcer.enable=true
+
+# Disables policy enforcement for a path
+quarkus.keycloak.policy-enforcer.paths.1.path=/public
+quarkus.keycloak.policy-enforcer.paths.1.enforcement-mode=DISABLED
+
+quarkus.health.extensions.enabled=true
+quarkus.smallrye-health.liveness-path=/public/health
+#quarkus.smallrye-health.readiness-path=/public/ready
+
 # Database
 quarkus.datasource.url=jdbc:postgresql:quarkus_test
 quarkus.datasource.driver=org.postgresql.Driver
 quarkus.datasource.username=quarkus_test
 quarkus.datasource.password=quarkus_test
-quarkus.datasource.max-size=8
-quarkus.datasource.min-size=2
 quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.log.sql=true
+quarkus.hibernate-orm.sql-load-script = import-test.sql
+
+# parametres applicatifs
+access.max.attempt=10
+access.recovery.attempt=5
+access.authentication.key=12345
+level.generation=5
