+++ src/test/resources/import-test.sql
@@ -1,0 +1,4 @@
+INSERT INTO pg_arbre_genealogie (id_chien,id_pere,id_mere) VALUES (1725763, 1725476, 1725231);
+INSERT INTO pg_data (iddog,data) VALUES (1725763, '{"id" : 1725763, "name" : "ELGA", "gender" : "F", "birthDate" : "1989-03-28", "birthCountry" : "FR", "inbreedingCoefficient" : "0%", "pedigrees" : [{   "country" : "FR",   "type" : "LOF",   "number" : "98993/17939" }], "tokens" : [{   "type" : "tatoo",   "number" : "JVP685" }], "breed" : {   "idRace" : 223,   "codeFci" : "002" }, "health" : [{   "code" : "HD-A",   "libelle" : "" }], "breeder" : {   "civility" : "M.",   "lastName" : "LONDEIX" }, "titles" : [{   "code" : "CHA",   "name" : "CHAMPION DES FIELD-TRIALS D''AUTOMNE" }, {   "code" : "CHCS",   "name" : "CHAMPION DE FRANCE DE CONFORMITE AU STANDARD" }, {   "code" : "TR",   "name" : "TRIALER" }]}');
+
+alter table pg_data alter column data type jsonb using data::JSON;
