+++ src/test/resources/quarkus-realm.json
@@ -465,7 +465,7 @@
           "attributes": {},
           "_id": "df1b74a9-3f10-499d-a581-368de48e512b",
           "uris": [
-            "/api/users/*"
+            "/api/pedigrees/*"
           ]
         },
         {
